add rarfile rar archive implementation
