+++ src/hoarder/archives/__init__.py
@@ -2,6 +2,8 @@ from .hash_archive import Algo, FileEntry, HashArchive
 from .hash_archive_repository import HashArchiveRepository
 from .hash_name_archive import HashNameArchive
 from .rar_archive import RarArchive
+from .rar_archive_7z import Rar7zArchive
+from .rar_archive_rarfile import RarfileRarArchive
 from .rar_path import RarScheme
 from .sfv_archive import SfvArchive
 
@@ -11,11 +13,15 @@ __all__ = [
     "HashArchive",
     "HashArchiveRepository",
     "HashNameArchive",
+    "Rar7zArchive",
     "RarArchive",
+    "RarfileRarArchive",
     "RarScheme",
     "SfvArchive",
     "hash_archive",
     "hash_archive_repository",
     "rar_archive",
+    "rar_archive_7z",
+    "rar_archive_rarfile",
     "sfv_archive",
 ]
