+++ tests/test_rar_file.py
@@ -3,7 +3,7 @@ import pathlib
 
 import pytest
 import tests.test_case_file_info
-from hoarder.archives import FileEntry, RarArchive, RarScheme
+from hoarder.archives import FileEntry, Rar7zArchive, RarScheme
 
 logger = logging.getLogger("hoarder.test_rar_file")
 
@@ -26,7 +26,7 @@ def test_rar_archives_set(
     main_archive_path = main_archive_path.resolve()
     root = main_archive_path.parent
     path = pathlib.PurePath(main_archive_path.name)
-    rar_archive = RarArchive.from_path(root, path, password=password)
+    rar_archive = Rar7zArchive.from_path(root, path, password=password)
     logger.debug(f"== Listing {main_archive_path}")
     for f in rar_archive.files:
         logger.debug(f)
